Fix issue on rbac snippet (#15)
* Adds tests for rbac step 5.

- Makes sure that we can't add new fields.
- That we can't remove fields.
- That we can't swap fields.

* Fixes an error on one of the snippets.

The code originally compared the resource.size of the modification
request vs the original. This actually is valid, as this property is
added. It doesn't work as expected, it doesn't ensure that the document
has been modified, but not have fields added.

- It compares the size of the document, which is going to be different
even when data was changed valid.
- Comparing size is not ideal either. Someone could remove a field and
add a new one to have this be valid. Instead comparing the keys of the
data directly ensures that all the fields are the same.